requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --extra-index-url=https://download.pytorch.org/whl/cu113 --output-file=./requirements/requirements.txt setup.py
#
--extra-index-url https://download.pytorch.org/whl/cu113

certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
idna==3.4
    # via requests
numpy==1.23.4
    # via
    #   gestrol (setup.py)
    #   opencv-python
    #   torchvision
opencv-python==4.6.0.66
    # via gestrol (setup.py)
pillow==9.2.0
    # via
    #   gestrol (setup.py)
    #   torchvision
requests==2.28.1
    # via torchvision
torch==1.11.0+cu113
    # via
    #   gestrol (setup.py)
    #   torchvision
torchvision==0.12.0+cu113
    # via gestrol (setup.py)
typing-extensions==4.4.0
    # via
    #   torch
    #   torchvision
urllib3==1.26.12
    # via requests
